+++ web/modeshape-web-jcr-rest/src/main/java/org/modeshape/web/jcr/rest/ServerHandler.java
@@ -1,8 +1,13 @@
 package org.modeshape.web.jcr.rest;
 
+import java.util.ArrayList;
+import java.util.List;
+import javax.jcr.Repository;
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
+import javax.jcr.Value;
 import javax.servlet.http.HttpServletRequest;
+import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
 import org.modeshape.common.annotation.Immutable;
@@ -57,13 +62,38 @@ class ServerHandler extends AbstractHandler {
                 JSONObject mapped = new JSONObject();
                 mapped.put("repository", repository);
                 jsonRepositories.put(name, mapped);
-                // } catch (RepositoryException e) {
-                // // Ignore, because we can't log in and thus cannot figure out any of the workspace names ...
-            } catch (Throwable t) {
-                t.printStackTrace();
+            } catch (RepositoryException e) {
+                // Ignore, because we can't log in and thus cannot figure out any of the workspace names ...
             }
         }
 
         return jsonRepositories.toString();
+    }
+
+    protected JSONObject getRepositoryMetadata( Session session ) throws JSONException, RepositoryException {
+        JSONObject metadata = new JSONObject();
+        Repository repository = session.getRepository();
+        for (String key : repository.getDescriptorKeys()) {
+            Value[] values = repository.getDescriptorValues(key);
+            if (values == null) continue;
+            if (values.length == 1) {
+                Value value = values[0];
+                if (value == null) continue;
+                metadata.put(key, jsonEncodedStringFor(value));
+            } else {
+                List<String> valueStrings = new ArrayList<String>();
+                for (Value value : values) {
+                    if (value == null) continue;
+                    valueStrings.add(jsonEncodedStringFor(value));
+                }
+                if (valueStrings.isEmpty()) continue;
+                if (valueStrings.size() == 1) {
+                    metadata.put(key, valueStrings.get(0));
+                } else {
+                    metadata.put(key, new JSONArray(valueStrings));
+                }
+            }
+        }
+        return metadata;
     }
 }
