--- web/modeshape-web-jcr-rest/src/main/java/org/modeshape/web/jcr/rest/model/RepositoryEntry.java
@@ -1,64 +1,0 @@
-/*
- * ModeShape (http://www.modeshape.org)
- * See the COPYRIGHT.txt file distributed with this work for information
- * regarding copyright ownership.  Some portions may be licensed
- * to Red Hat, Inc. under one or more contributor license agreements.
- * See the AUTHORS.txt file in the distribution for a full listing of 
- * individual contributors. 
- *
- * ModeShape is free software. Unless otherwise indicated, all code in ModeShape
- * is licensed to you under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * ModeShape is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package org.modeshape.web.jcr.rest.model;
-
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-@Deprecated
-@XmlRootElement( name = "repository" )
-public class RepositoryEntry {
-
-    private String name;
-    private RepositoryResources resources;
-
-    public RepositoryEntry() {
-        resources = new RepositoryResources();
-    }
-
-    public RepositoryEntry( String contextName,
-                            String repositoryName ) {
-        this.name = repositoryName;
-
-        resources = new RepositoryResources(contextName, repositoryName);
-    }
-
-    @XmlElement
-    public String getName() {
-        return name;
-    }
-
-    public void setName( String name ) {
-        this.name = name;
-    }
-
-    @XmlElement
-    public RepositoryResources getResources() {
-        return resources;
-    }
-
-    public void setResources( RepositoryResources resources ) {
-        this.resources = resources;
-    }
-}
